Adds a basic User entity and resource to know the logged user
Adds a basic user with attribute role, login and password. This attributes are
validated. Also adds a resource to return a role if the user are logged.
Also is modified the archive web.xml in the layer rest to add a basic security.
In addition, adds SQL scripts for creating the user table and adding some data.