Adds the target number of sign types to detect
The target column has been added to the signtypes table to indicate the number
of signs to be achieved in order to satisfy the goal of sign types detected. In
addition, the list of signs identified by an user has been refactored to
include it as a filter in the list of signs.