Adds a logical deletion mechanism to the different entities
Adds two fields to the entities that represent if the data has been deleted and
when. This mechanism allows implementing a logical deletion of the database in
order to perform a transparent deletion to the user but internally keeping the
deleted information. In this way, if necessary, the deletion action can be
reversed.