Defines two new sign types
This commit adds as type of signs: "not finding" and " non-normal". The first
type of sign refers to those radiographs that have a healthy appearance and no
signs have been detected. On the other hand, the second type of sign is the
detection of signs not defined in the list of available signs. Both types of
signs have not an associated location.