Adds the functionality to create an exploration
Adds the resource which manages the creation of an exploration. To this end, it
creates the classes in order to transform the request into the elements that
compose an exploration. In addition, the request size that the Wildfly
application server can receive is increased from 10MB to 100MB. This change is
made because when creating an exploration, it can contain two radiograph images
which exceeded the server's default size limit.