Adds functionality to save and retrieve radiographs on disk
A specific resource has been added to retrieve the bytes of a radiograph. In
addition, the resource for obtaining a radiograph has been adapted to provide
only its metadata without the image representation. On the other hand, a path
has been configured on the server where the radiographs of the explorations
performed are stored. Finally, the 'commons-fileupload' dependency has been
added to facilitate the management of the storage and recuperation of files on
disk.