Adds functionality to edit an existing exploration
A new resource has been added to allow editing the exploration from an id.
Before editing an exploration, a backup copy of the radiographs saved on disk
is made to avoid loss of information and to ensure that if the latest version
is not saved correctly, the previous version can be restored.