Implements an access control list (ACL)
A permission mechanism is added based on the definition of the available
functionalities and their enabled actions. This functionality-action
combination will be associated to a role/group. Every user will have a unique
role associated to it, having the assigned permissions and privileges. To this
end, the classes and resources needed to implement the ACL have been added
including the permission conditions associated with each of the application's
functionalities.